chore: cleanup changesetList
