+++ src/utils/filters.js
@@ -1,4 +1,7 @@
+// @flow
 import { List, Map } from 'immutable';
+import moment from 'moment';
+import { DEFAULT_FROM_DATE } from '../config/constants';
 
 export function validateFilters(filters: Map<string, *>): boolean {
   if (!Map.isMap(filters)) return false;
@@ -16,4 +19,16 @@ export function validateFilters(filters: Map<string, *>): boolean {
     }
   });
   return valid;
+}
+
+export function getDefaultFromDate(): Array<Object> {
+  const defaultDate = moment()
+    .subtract(DEFAULT_FROM_DATE, 'days')
+    .format('YYYY-MM-DD');
+  return [
+    {
+      label: defaultDate,
+      value: defaultDate
+    }
+  ];
 }
